chore: temporary extract tests as file
